refactor instructiosn
